move the add_x_binding functions from config
They are now in a newly added binding object that will hold all these
functions until I come up with a better way of adding hotkeys. This also
adds a couple of missing function prototypes for the more recent
additions to these functions that did not error because of convinient
placement of the definitions.